refactor Theme into `components/foundations`
