+++ src/components/foundations/index.ts
@@ -1,3 +1,4 @@
 export * from './box';
 export * from './reset';
 export * from './typography';
+export * from './Theme';
